Implement Interactive CLI (#357)
* * renamed cli to wallet
* use conf file instead of account/ committee files and arg values
* removed create accounts command
* added add-address command

* removed benchmark from CLI commands

* revert client.rs code

* more refactoring
* got structopt working with shellfish

add sync client command

added transfer command

added objects command

cli refactoring

* get_owned_objects don't need to be async

* fixup after rebase

* * port should be u16
* added fastx binary for genesis and start authorities

* * more verbose

* * remove portpicker
* new PortAllocator

* * move call using args instead of config file

* * minor improvements

* * remove shellfish
* command completion using rustyline
* new Config trait for easy loading/storing config files
* address some PR comment

* fixup after rebase

* replace rustyline with reedline to avoid BSL license

* fixup after rebase

* * command completion support multiple command flag

* Revert "replace rustyline with reedline to avoid BSL license"

This reverts commit 371dce41

* * fixup after revert
* allow BSL-1.0 license

* * allow BSL-1.0 license

* * add back clear command

* * bugfix genesis not writing to wallet.conf

* fixup after rebase

* use unescape crate to handle unescaping string
address PR comments

* fixup after rebase

+++ fastpay_core/src/authority_server.rs
@@ -19,7 +19,7 @@ pub struct AuthorityServer {
 impl AuthorityServer {
     pub fn new(
         base_address: String,
-        base_port: u32,
+        base_port: u16,
         buffer_size: usize,
         state: AuthorityState,
     ) -> Self {
